Revert "adding benchmarking logic"
This reverts commit 4b44895f499fc122ae7263cc58cb8be67a212a6d.
