+++ circom-proving/Cargo.toml
@@ -28,3 +28,4 @@ ark-crypto-primitives = { version = "^0.3.0", features = ["r1cs"], default-featu
 
 [features]
 default = ["wasmer/sys-default"]
+
