+++ pallets/anonymity-mining-claims/Cargo.toml
@@ -4,7 +4,7 @@ description = "Pallet that handles anonymity mining claims."
 edition = "2018"
 homepage = "https://substrate.dev"
 license = "Unlicense"
-name = "pallet-anonimity-mining-claims"
+name = "pallet-anonymity-mining-claims"
 repository = "https://github.com/webb-tools/protocol-substrate"
 version = "1.0.0"
 
@@ -24,42 +24,41 @@ pallet-signature-bridge = { path = "../../pallets/signature-bridge", default-fea
 pallet-token-wrapper = { path = "../token-wrapper", default-features = false }
 pallet-vanchor = { path = "../vanchor", default-features = false }
 pallet-vanchor-handler = { path = "../../pallets/vanchor-handler", default-features = false }
-pallet-claims-verifier = { path = "../claims-verifier", default-features = false }
-pallet-hasher = { path = "../hasher", default-features = false }
 
 frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 frame-system = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
-sp-io = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 webb-primitives = { path = "../../primitives", default-features = false }
-ark-std = { version = "^0.3.0", default-features = false }
-ark-bn254 = { version = "^0.3.0", default-features = false, features = ["curve"] }
-ark-ff = { version = "^0.3.0", default-features = false }
-ark-serialize = { version = "^0.3.0", default-features = false, features = ["derive"] }
-circom-proving = { path = "../../circom-proving", default-features = false }
-frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
-arkworks-setups = { version = "1.2.1", features = ["r1cs"], default-features = false }
-ark-circom = { git = "https://github.com/vacp2p/ark-circom", branch = "wasm", default-features = false, features = ["circom-2"] }
-hex = "0.4"
-orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
-ark-relations = { version = "^0.3.0", default-features = false }
-ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f", default-features = false, features = ["parallel"] }
 
 [dev-dependencies]
 ark-bls12-381 = { version = "^0.3.0", default-features = false, features = ["curve"] }
+ark-bn254 = { version = "^0.3.0", default-features = false, features = ["curve"] }
 ark-crypto-primitives = { version = "^0.3.0", features = ["r1cs"], default-features = false }
 ark-ec = { version = "^0.3.0", default-features = false }
+ark-ff = { version = "^0.3.0", default-features = false }
+ark-relations = { version = "^0.3.0", default-features = false }
+ark-serialize = { version = "^0.3.0", default-features = false, features = ["derive"] }
+ark-std = { version = "^0.3.0", default-features = false }
+arkworks-setups = { version = "1.2.1", features = ["r1cs"], default-features = false }
+frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
+hex = "0.4"
 orml-currencies = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
+orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
+pallet-hasher = { path = "../hasher", default-features = false }
 pallet-key-storage = { path = "../key-storage" }
 pallet-timestamp = { default-features = true, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.36" }
 pallet-vanchor-verifier = { path = "../vanchor-verifier", default-features = false }
+pallet-claims-verifier = { path = "../claims-verifier", default-features = false }
 serde = { version = "1.0.119" }
 sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
+sp-io = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 
 # For ark-circom
 circom-proving = { path = "../../circom-proving", default-features = false }
+ark-groth16 = { git = "https://github.com/arkworks-rs/groth16", rev = "765817f", default-features = false, features = ["parallel"] }
+ark-circom = { git = "https://github.com/vacp2p/ark-circom", branch = "wasm", default-features = false, features = ["circom-2"] }
 num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
 once_cell = "1.14.0"
 cfg-if = "1.0"
@@ -95,3 +94,4 @@ std = [
   "pallet-token-wrapper/std",
 
 ]
+wasm = ["wasmer/js", "wasmer/std"]
