+++ pallets/claims-verifier/src/benchmarking.rs
@@ -45,9 +45,9 @@ benchmarks_instance_pallet! {
 		let c in 0..MAX_VERIFIER_LENGTH;
 		let depositor: T::AccountId = account("depositor", 0, SEED);
 		let parameters = vec![0u8;c as usize];
-	}: _(RawOrigin::Root, 1u8, parameters.clone().try_into().unwrap())
+	}: _(RawOrigin::Root, (1u8,1u8), parameters.clone().try_into().unwrap())
 	verify {
-		assert_eq!(Pallet::<T, I>::parameters(1u8), parameters);
+		assert_eq!(Pallet::<T, I>::parameters((1u8,1u8)), parameters);
 	}
 }
