+++ pallets/vanchor/Cargo.toml
@@ -22,15 +22,10 @@ pallet-linkable-tree = { path = "../linkable-tree", default-features = false }
 pallet-mt = { path = "../mt", default-features = false }
 webb-primitives = { path = "../../primitives", default-features = false, features = ["hashing", "verifying", "field_ops"] }
 orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
-orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
 pallet-asset-registry = { path = "../asset-registry", default-features = false }
 pallet-token-wrapper = { path = "../token-wrapper", default-features = false }
-pallet-hasher = { path = "../hasher", default-features = false }
-pallet-vanchor-verifier = { path = "../vanchor-verifier", default-features = false }
 frame-system-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36", optional = true }
-frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 arkworks-native-gadgets = { version = "1.2.0", default-features = false }
-arkworks-setups = { version = "1.2.1", features = ["r1cs"], default-features = false }
 ark-std = { version = "^0.3.0", default-features = false }
 ark-bn254 = { version = "^0.3.0", default-features = false, features = [ "curve" ] }
 ark-ff = { version = "^0.3.0", default-features = false}
@@ -45,8 +40,12 @@ sp-runtime = { default-features = false, git = "https://github.com/paritytech/su
 pallet-balances = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 pallet-treasury = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 pallet-key-storage = {path = "../key-storage"}
+pallet-hasher = { path = "../hasher", default-features = false }
+pallet-vanchor-verifier = { path = "../vanchor-verifier", default-features = false }
 circom-proving = { path = "../../circom-proving", default-features = false }
+orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
 orml-currencies = { git = "https://github.com/open-web3-stack/open-runtime-module-library.git", branch = "polkadot-v0.9.36", default-features = false }
+arkworks-setups = { version = "1.2.1", features = ["r1cs"], default-features = false }
 ark-crypto-primitives = { version = "^0.3.0", features = ["r1cs"], default-features = false }
 ark-ff = { version = "0.3.0", default-features = false, features = [ "asm"] }
 ark-ec = { version = "^0.3.0", default-features = false }
@@ -57,6 +56,7 @@ ark-std = { version = "^0.3.0", default-features = false }
 ark-bn254 = { version = "0.3.0" }
 ark-circom = { git = "https://github.com/vacp2p/ark-circom", branch = "wasm", default-features = false, features = ["circom-2"] }
 num-bigint = { version = "0.4", default-features = false, features = ["rand"] }
+frame-benchmarking = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36" }
 # For ark-circom
 once_cell = "1.14.0"
 cfg-if = "1.0"
