+++ pallets/vanchor/src/lib.rs
@@ -48,8 +48,8 @@ mod test_utils;
 mod tests;
 #[cfg(test)]
 mod tests_circom;
-// #[cfg(test)]
-// mod zerokit_utils;
+#[cfg(test)]
+mod zerokit_utils;
 
 mod benchmarking;
 mod benchmarking_utils;
