+++ standalone/runtime/Cargo.toml
@@ -41,8 +41,6 @@ pallet-token-wrapper-handler = { path = "../../pallets/token-wrapper-handler", d
 pallet-vanchor = { path = "../../pallets/vanchor", default-features = false }
 pallet-vanchor-handler = { path = "../../pallets/vanchor-handler", default-features = false }
 pallet-vanchor-verifier = { path = "../../pallets/vanchor-verifier", default-features = false }
-pallet-claims-verifier = { path = "../../pallets/claims-verifier", default-features = false }
-pallet-anonimity-mining-claims = { path = "../../pallets/anonymity-mining-claims", default-features = false }
 pallet-verifier = { path = "../../pallets/verifier", default-features = false }
 webb-primitives = { path = "../../primitives", default-features = false, features = ["hashing"] }
 
@@ -135,7 +133,6 @@ runtime-benchmarks = [
   "pallet-token-wrapper-handler/runtime-benchmarks",
   "pallet-relayer-registry/runtime-benchmarks",
   "pallet-key-storage/runtime-benchmarks",
-  "pallet-anonimity-mining-claims/runtime-benchmarks",
 ]
 
 std = [
@@ -196,8 +193,6 @@ std = [
   "pallet-hasher/std",
   "pallet-verifier/std",
   "pallet-vanchor-verifier/std",
-  "pallet-claims-verifier/std",
-  "pallet-anonimity-mining-claims/std",
   "pallet-mt/std",
   "pallet-linkable-tree/std",
   "pallet-mixer/std",
